refactor: delete dead EffortStatusConfig helpers + extract EFFORT_STATUS_OPTIONS (RFC 31c1a0be D2a, #3194)
D2a of the Phase 4 stage-5 enum cascade cleanup. Audit before edits found
that of the 9 exports in `EffortStatusConfig.ts`, only `EFFORT_STATUS_OPTIONS`
has any production consumer (single file: `StatusSelectPropertyField.ts`).
The remaining 8 — `EFFORT_STATUS_CONFIG`, `STATUS_NAME_TO_ENUM`,
`STATUS_NAME_TO_WIKILINK`, `normalizeEffortStatus`, `isDoneStatus`,
`isTrashedStatus`, `getEffortStatusLabel`, and `type EffortStatusName` —
are referenced only by the file's own test suite and the two
re-export barrels. `test-utils` re-defines `EffortStatusName`,
`isDoneStatus`, `isTrashedStatus` independently in
`packages/test-utils/src/{types,helpers/status.helpers}.ts`, so deletion is
non-coupling for the test factories.

## Changes

- New `packages/exocortex/src/domain/constants/EffortStatusOptions.ts` —
  holds only `EFFORT_STATUS_OPTIONS`. Symbolic wikilink values
  (`"[[ems__EffortStatusDoing]]"`) preserved verbatim: changing them
  requires the coordinated SPARQLTemplateLibrary + NoteToRDFConverter +
  starter-kit ASK migration (D2c+D2d, descoped — see below).
- Delete `EffortStatusConfig.ts` (143 LOC, 8 of 9 exports dead).
- Delete `EffortStatusConfig.test.ts` (135 LOC of dead-helper coverage).
- New `EffortStatusOptions.test.ts` (preserves the two `EFFORT_STATUS_OPTIONS`
  shape/value-pair assertions from the legacy suite).
- Update both re-export barrels (`packages/exocortex/src/index.ts`,
  `packages/exocortex/src/domain/constants/index.ts`) to drop the dead
  symbols and surface only `EFFORT_STATUS_OPTIONS`.
- Refresh the directional doc note on `EffortStatus.ts` to (a) reflect the
  post-PR-D cascade math (~25 files, 18+ direct enum call sites; well above
  the 15-caller cascade-cap), (b) add the starter-kit ASK preconditions to
  the explicit list of coupled surfaces, and (c) point at the new
  `EffortStatusOptions.ts` home for the only live export.

## Caller audit

| Export | Prod callers | Test callers |
|---|---|---|
| `EFFORT_STATUS_CONFIG` | 0 | 1 (this file's test) |
| `EffortStatusName` (type) | 0 | 1 (test-utils has its own) |
| `STATUS_NAME_TO_ENUM` | 0 | 1 |
| `STATUS_NAME_TO_WIKILINK` | 0 | 1 |
| `EFFORT_STATUS_OPTIONS` | 1 (StatusSelectPropertyField) | 2 |
| `normalizeEffortStatus` | 0 | 1 |
| `isDoneStatus` | 0 | 1 (test-utils has its own copy) |
| `isTrashedStatus` | 0 | 1 (test-utils has its own copy) |
| `getEffortStatusLabel` | 0 | 1 |

## D2b, D2c, D2d descope (rationale)

The companion sub-PRs from the D2 prompt are descoped here rather than
shipped as separate PRs because Phase 4 PR-B (#3202) already documented
the descope reasoning verbatim in the JSDoc invariants on the affected
files; no additional artefact is needed.

- **D2b — `EffortStatus.ts` enum deletion.** 18+ direct enum call sites
  across `WorkflowEngine`, `WorkflowResolver`, `WorkflowCommandAdapter`,
  `VisibilityGenerator`, `EffortVisibilityRules`, `ProjectVisibilityRules`,
  `DefaultWorkflows`, `WorkflowDefinition`, `KanbanColumnProvider`,
  `DailyTasksRenderer`, `BlockerHelpers`, `cli/workflow.ts`, plus matching
  test files — well above the 15-caller cascade-cap. JSDoc directional note
  is already in place (added in PR-B #3202 at lines 1-26 of
  `EffortStatus.ts`); this PR refreshes the cascade math to reflect the
  post-D2a state.

- **D2c — `SPARQLTemplateLibrary.EFFORT_STATUSES` migration.** Tight
  coupling with `NoteToRDFConverter.expandClassValue`: both co-emit the
  same canonical class IRI (`<https://exocortex.my/ontology/ems#EffortStatusDoing>`).
  Changing either alone silently breaks the template queries (Archive
  Completed, Doing Tasks). Status predicates have no dual-storage
  fallback — `isProtoPredicate` (NoteToRDFConverter:998-1004) restricts
  dual-storage to `exo__Asset_prototype` only. The PR-B JSDoc invariant on
  `SPARQLTemplateLibrary.ts` already documents this coupling.

- **D2d — `NoteToRDFConverter.expandClassValue` migration.** Same
  coupling as D2c, plus a cross-repo dependency: the starter-kit's 14
  status-command `exocmd__Precondition_sparqlAsk` filters (e.g.
  `/exocmd/status/656d075c-...md`) match the symbolic
  `<https://exocortex.my/ontology/ems#EffortStatusToDo>` IRI emitted here.
  Removing the symbolic-IRI emission branch without a coordinated
  starter-kit PR regresses every status-transition button visibility
  rule. The PR-B JSDoc invariant at `NoteToRDFConverter.valueToRDFObject`
  lines 980-983 (origin/main) already documents this. Phase 3 hotfix
  bypass for `Grounding_targetValueRef` / `Grounding_targetValueSubstitution`
  (PR #3199, b79c7ae4) is preserved verbatim — not touched.

D2c+D2d collectively require a cross-repo coordinated PR series
(exocortex `EFFORT_STATUSES` + `expandClassValue` + starter-kit ASK
preconditions, all three together). This is RFC 31c1a0be Phase 5b
territory and beyond the cascade-cap for any single sub-PR in this stage.

## Test plan

- [x] `npm run check:types` → clean
- [x] `npm run lint` → 2 pre-existing warnings (`DisplayNameResolver:77`,
      `PrintNameRuleService:167`), 0 errors
- [x] `npx jest packages/exocortex/tests/domain/constants` → 4/4 suites,
      54/54 tests pass
- [x] `npx jest packages/obsidian-plugin/tests/unit/property-fields/StatusSelectPropertyField.test.ts`
      → 20/20 pass
- [x] `npx jest packages/obsidian-plugin/tests/unit` → 238/238 suites,
      4912/4915 pass (3 pre-existing skips)
- [x] `npx jest packages/exocortex/tests/unit packages/exocortex/tests/domain`
      → 175/180 suites pass. 5 failures verified pre-existing on
      origin/main (1eb523c4) by `git stash`-baseline comparison:
      `CommandConstants.test.ts` (GroundingType length 9 vs expected 6 —
      stale assertion after extension), `CommandDefinition.test.ts`
      (`targetFolder` property no longer exists on grounding type after
      RFC 31c1a0be Phase 5a removal), plus `QueryExecutor.branch`,
      `sparql/subquery-acceptance`, and `FrontmatterService.error`.
