+++ packages/exocortex/src/domain/constants/index.ts
@@ -1,16 +1,6 @@
 export { AssetClass } from "./AssetClass";
 export { EffortStatus } from "./EffortStatus";
-export {
-  type EffortStatusName,
-  EFFORT_STATUS_CONFIG,
-  STATUS_NAME_TO_ENUM,
-  STATUS_NAME_TO_WIKILINK,
-  EFFORT_STATUS_OPTIONS,
-  normalizeEffortStatus,
-  isDoneStatus,
-  isTrashedStatus,
-  getEffortStatusLabel,
-} from "./EffortStatusConfig";
+export { EFFORT_STATUS_OPTIONS } from "./EffortStatusOptions";
 export { GroundingType } from "./GroundingType";
 export {
   CommandProperty,
